added some code to entity

+++ protocol/my_ctype/my_ctypes.c
@@ -80,7 +80,7 @@ void pointerFill(reg n, u8* ptr, u8 data)
 
 
 // string compleate------------------------------------------
-u8 cTypeStrnCmp(reg n, const c8* str1, const c8* str2)
+int cTypeStrnCmp(reg n, const c8* str1, const c8* str2)
 {
     while(n--) {
         if(*str1++ != *str2++) {
@@ -90,7 +90,7 @@ u8 cTypeStrnCmp(reg n, const c8* str1, const c8* str2)
     return 0;
 }
 
-u8 cTypeStrnRevCmp(reg n, const c8* str1, const c8* str2)
+int cTypeStrnRevCmp(reg n, const c8* str1, const c8* str2)
 {
     while(n--) {
         if(*str1++ != *(str2 + n)) {
